spostato il vecchio comando deb in man e creato un nuovo comando deb
